macOS support fixes

+++ src/util/net/fd_ip4.c
@@ -19,6 +19,6 @@ fd_cstr_to_ip4_addr( char const * s,
   x[ 2 ] = fd_cstr_to_uchar( tok[ 2 ] );
   x[ 3 ] = fd_cstr_to_uchar( tok[ 3 ] );
 
-  *out = ( x[3] | (x[2]<<8) | (x[1]<<16) | (x[0]<<24) );
+  *out = ( (uint)x[3] | ((uint)x[2]<<8) | ((uint)x[1]<<16) | ((uint)x[0]<<24) );
   return 1;
 }
